+++ tests/integration/distribution/keeper/msg_server_test.go
@@ -194,10 +194,10 @@ func TestMsgWithdrawDelegatorReward(t *testing.T) {
 	// setup delegation
 	delTokens := sdk.TokensFromConsensusPower(2, sdk.DefaultPowerReduction)
 	validator, issuedShares := validator.AddTokensFromDel(delTokens)
-	delegation := stakingtypes.NewDelegation(delAddr, validator.GetOperator(), issuedShares)
-	assert.NilError(t, f.stakingKeeper.SetDelegation(f.sdkCtx, delegation))
-	err = f.distrKeeper.SetDelegatorStartingInfo(f.sdkCtx, validator.GetOperator(), delAddr, distrtypes.NewDelegatorStartingInfo(2, math.LegacyOneDec(), 20))
-	require.NoError(t, err)
+
+	delegation := stakingtypes.NewDelegation(delAddr.String(), validator.GetOperator().String(), issuedShares)
+	require.NoError(t, f.stakingKeeper.SetDelegation(f.sdkCtx, delegation))
+	require.NoError(t, f.distrKeeper.SetDelegatorStartingInfo(f.sdkCtx, validator.GetOperator(), delAddr, distrtypes.NewDelegatorStartingInfo(2, math.LegacyOneDec(), 20)))
 	// setup validator rewards
 	decCoins := sdk.DecCoins{sdk.NewDecCoinFromDec(sdk.DefaultBondDenom, math.LegacyOneDec())}
 	historicalRewards := distrtypes.NewValidatorHistoricalRewards(decCoins, 2)