+++ tests/integration/staking/keeper/validator_bench_test.go
@@ -1,6 +1,7 @@
 package keeper_test
 
 import (
+	"bytes"
 	"fmt"
 	"testing"
 
@@ -59,8 +60,11 @@ func BenchmarkGetValidatorDelegations(b *testing.B) {
 			delegator := sdk.AccAddress(fmt.Sprintf("address%d", i))
 			banktestutil.FundAccount(f.sdkCtx, f.bankKeeper, delegator,
 				sdk.NewCoins(sdk.NewCoin(sdk.DefaultBondDenom, math.NewInt(int64(i)))))
-			NewDel := types.NewDelegation(delegator, val, math.LegacyNewDec(int64(i)))
-			f.stakingKeeper.SetDelegation(f.sdkCtx, NewDel)
+			NewDel := types.NewDelegation(delegator.String(), val.String(), math.LegacyNewDec(int64(i)))
+
+			if err := f.stakingKeeper.SetDelegation(f.sdkCtx, NewDel); err != nil {
+				panic(err)
+			}
 		}
 	}
 
@@ -90,10 +94,11 @@ func BenchmarkGetValidatorDelegationsLegacy(b *testing.B) {
 	for _, val := range valAddrs {
 		for i := 0; i < delegationsNum; i++ {
 			delegator := sdk.AccAddress(fmt.Sprintf("address%d", i))
-			banktestutil.FundAccount(f.sdkCtx, f.bankKeeper, delegator,
-				sdk.NewCoins(sdk.NewCoin(sdk.DefaultBondDenom, math.NewInt(int64(i)))))
-			NewDel := types.NewDelegation(delegator, val, math.LegacyNewDec(int64(i)))
-			f.stakingKeeper.SetDelegation(f.sdkCtx, NewDel)
+			banktestutil.FundAccount(f.sdkCtx, f.bankKeeper, delegator, sdk.NewCoins(sdk.NewCoin(sdk.DefaultBondDenom, math.NewInt(int64(i)))))
+			NewDel := types.NewDelegation(delegator.String(), val.String(), math.LegacyNewDec(int64(i)))
+			if err := f.stakingKeeper.SetDelegation(f.sdkCtx, NewDel); err != nil {
+				panic(err)
+			}
 		}
 	}
 
@@ -114,8 +119,15 @@ func updateValidatorDelegationsLegacy(f *fixture, existingValAddr, newValAddr sd
 
 	for ; iterator.Valid(); iterator.Next() {
 		delegation := types.MustUnmarshalDelegation(cdc, iterator.Value())
-		if delegation.GetValidatorAddr().Equals(existingValAddr) {
-			k.RemoveDelegation(f.sdkCtx, delegation)
+		valAddr, err := k.ValidatorAddressCodec().StringToBytes(delegation.GetValidatorAddr())
+		if err != nil {
+			panic(err)
+		}
+
+		if bytes.EqualFold(valAddr, existingValAddr) {
+			if err := k.RemoveDelegation(f.sdkCtx, delegation); err != nil {
+				panic(err)
+			}
 			delegation.ValidatorAddress = newValAddr.String()
 			k.SetDelegation(f.sdkCtx, delegation)
 		}