+++ x/slashing/keeper/msg_server_test.go
@@ -214,7 +214,7 @@ func (s *KeeperTestSuite) TestUnjail() {
 				s.slashingKeeper.SetValidatorSigningInfo(s.ctx, sdk.ConsAddress(addr), info)
 
 				s.stakingKeeper.EXPECT().Validator(s.ctx, valAddr).Return(val, nil)
-				del := types.NewDelegation(addr, valAddr, sdkmath.LegacyNewDec(100))
+				del := types.NewDelegation(addr.String(), valAddr.String(), sdkmath.LegacyNewDec(100))
 
 				s.stakingKeeper.EXPECT().Delegation(s.ctx, addr, valAddr).Return(del, nil)
 
@@ -244,7 +244,7 @@ func (s *KeeperTestSuite) TestUnjail() {
 				s.slashingKeeper.SetValidatorSigningInfo(s.ctx, sdk.ConsAddress(addr), info)
 
 				s.stakingKeeper.EXPECT().Validator(s.ctx, valAddr).Return(val, nil)
-				del := types.NewDelegation(addr, valAddr, sdkmath.LegacyNewDec(100))
+				del := types.NewDelegation(addr.String(), valAddr.String(), sdkmath.LegacyNewDec(100))
 
 				s.stakingKeeper.EXPECT().Delegation(s.ctx, addr, valAddr).Return(del, nil)
 
@@ -274,7 +274,7 @@ func (s *KeeperTestSuite) TestUnjail() {
 				s.slashingKeeper.SetValidatorSigningInfo(s.ctx, sdk.ConsAddress(addr), info)
 
 				s.stakingKeeper.EXPECT().Validator(s.ctx, valAddr).Return(val, nil)
-				del := types.NewDelegation(addr, valAddr, sdkmath.LegacyNewDec(10000))
+				del := types.NewDelegation(addr.String(), valAddr.String(), sdkmath.LegacyNewDec(10000))
 
 				s.stakingKeeper.EXPECT().Delegation(s.ctx, addr, valAddr).Return(del, nil)
 
@@ -304,7 +304,7 @@ func (s *KeeperTestSuite) TestUnjail() {
 				s.slashingKeeper.SetValidatorSigningInfo(s.ctx, sdk.ConsAddress(addr), info)
 
 				s.stakingKeeper.EXPECT().Validator(s.ctx, valAddr).Return(val, nil)
-				del := types.NewDelegation(addr, valAddr, sdkmath.LegacyNewDec(100))
+				del := types.NewDelegation(addr.String(), valAddr.String(), sdkmath.LegacyNewDec(100))
 
 				s.stakingKeeper.EXPECT().Delegation(s.ctx, addr, valAddr).Return(del, nil)
 				s.stakingKeeper.EXPECT().Unjail(s.ctx, sdk.ConsAddress(addr)).Return(nil)